first attempt at hybrid solver setup
This doesn't work in its current form: the non-incremental solver is reset on
each run, but apparently that `reset` does not fully reset the solver. In
particular, on a second query to the non-incremental solver, a `define-fun` does
not lead to output `success` if there was another `define-fun` of the same
function before the `reset` (unlike the behaviour for datatype declarations,
etc).